requirements.txt
azure-functions

sqlalchemy

pydantic

typing-extensions

pyodbc

azure-identity